backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
neo4j==5.14.1
python-dotenv==1.0.0
marshmallow==3.20.1
uuid==1.30
pytest==7.4.3
